split code for x86 and aarch64 support
