+++ src/x86_64/mod.rs
@@ -1,185 +1,263 @@
-/*
-Copyright (c) 2016 Saurav Sachidanand
+pub mod consts;
+pub mod ffi;
 
-Permission is hereby granted, free of charge, to any person obtaining a copy
-of this software and associated documentation files (the "Software"), to deal
-in the Software without restriction, including without limitation the rights
-to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-copies of the Software, and to permit persons to whom the Software is
-furnished to do so, subject to the following conditions:
-
-The above copyright notice and this permission notice shall be included in
-all copies or substantial portions of the Software.
-
-THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
-THE SOFTWARE.
-*/
-
-//! Bindings to the Hypervisor Framework
-
+use self::ffi::*;
+use crate::{match_MemPerm, match_error_code, Error, MemPerm};
+use core::fmt;
 use libc::*;
 
-/// Hypervisor Framework return code
-pub type hv_return_t = u32;
-
-// Hypervisor Framework return codes
-pub const HV_SUCCESS: hv_return_t = 0;
-pub const HV_ERROR: hv_return_t = 0xfae94001;
-pub const HV_BUSY: hv_return_t = 0xfae94002;
-pub const HV_BAD_ARGUMENT: hv_return_t = 0xfae94003;
-pub const HV_NO_RESOURCES: hv_return_t = 0xfae94005;
-pub const HV_NO_DEVICE: hv_return_t = 0xfae94006;
-pub const HV_UNSUPPORTED: hv_return_t = 0xfae9400f;
-
-/// Options for hv_vcpu_create()
-pub type hv_vm_options_t = u64;
-pub const HV_VM_DEFAULT: hv_vm_options_t = 0 << 0;
-
-// Creating and Destroying VM Instances
-extern "C" {
-	/// Creates a VM instance for the current Mach task
-	pub fn hv_vm_create(flags: hv_vm_options_t) -> hv_return_t;
-
-	/// Destroys the VM instance associated with the current Mach task
-	pub fn hv_vm_destroy() -> hv_return_t;
+/// Creates a VM instance for the current Mach task
+pub fn create_vm() -> Result<(), Error> {
+	match_error_code(unsafe { hv_vm_create(HV_VM_DEFAULT) })
 }
 
-/// Type of a vCPU ID
-pub type hv_vcpuid_t = c_uint;
-
-// Option for hv_vcpu_create()
-pub const HV_VCPU_DEFAULT: u64 = 0;
-
-// Creating and Managing vCPU Instances
-extern "C" {
-	/// Creates a vCPU instance for the current thread
-	pub fn hv_vcpu_create(vcpu: *mut hv_vcpuid_t, flags: hv_vm_options_t) -> hv_return_t;
-
-	/// Executes a vCPU
-	pub fn hv_vcpu_run(vcpu: hv_vcpuid_t) -> hv_return_t;
-
-	/// Forces an immediate VMEXIT of a set of vCPUs of the VM
-	pub fn hv_vcpu_interrupt(vcpu: *const hv_vcpuid_t, vcpu_count: c_uint) -> hv_return_t;
-
-	/// Returns the cumulative execution time of a vCPU in nanoseconds
-	pub fn hv_vcpu_get_exec_time(vcpu: hv_vcpuid_t, time: *mut u64) -> hv_return_t;
-
-	/// Forces flushing of cached vCPU state
-	pub fn hv_vcpu_flush(vcpu: hv_vcpuid_t) -> hv_return_t;
-
-	/// Invalidates the TLB of a vCPU
-	pub fn hv_vcpu_invalidate_tlb(vcpu: hv_vcpuid_t) -> hv_return_t;
-
-	/// Destroys the vCPU instance associated with the current thread
-	pub fn hv_vcpu_destroy(vcpu: hv_vcpuid_t) -> hv_return_t;
+/// Maps a region in the virtual address space of the current task into the guest physical
+/// address space of the virutal machine
+pub fn map_mem(mem: &[u8], gpa: u64, mem_perm: &MemPerm) -> Result<(), Error> {
+	match_error_code(unsafe {
+		hv_vm_map(
+			mem.as_ptr() as *const c_void,
+			gpa as hv_gpaddr_t,
+			mem.len() as size_t,
+			match_MemPerm(mem_perm),
+		)
+	})
 }
 
-// Accessing Registers
-extern "C" {
-	/// Returns the current value of an architectural x86 register
-	/// of a vCPU
-	pub fn hv_vcpu_read_register(
-		vcpu: hv_vcpuid_t,
-		reg: super::Register,
-		value: *mut u64,
-	) -> hv_return_t;
-
-	/// Sets the value of an architectural x86 register of a vCPU
-	pub fn hv_vcpu_write_register(
-		vcpu: hv_vcpuid_t,
-		reg: super::Register,
-		value: u64,
-	) -> hv_return_t;
+/// Modifies the permissions of a region in the guest physical address space of the virtual
+/// machine
+pub fn protect_mem(gpa: u64, size: usize, mem_perm: &MemPerm) -> Result<(), Error> {
+	match_error_code(unsafe {
+		hv_vm_protect(gpa as hv_gpaddr_t, size as size_t, match_MemPerm(mem_perm))
+	})
 }
 
-// Accessing Floating Point (FP) State
-extern "C" {
-	/// Returns the current architectural x86 floating point and
-	/// SIMD state of a vCPU
-	pub fn hv_vcpu_read_fpstate(
-		vcpu: hv_vcpuid_t,
-		buffer: *mut c_void,
-		size: size_t,
-	) -> hv_return_t;
-
-	/// Sets the architectural x86 floating point and SIMD state of
-	/// a vCPU
-	pub fn hv_vcpu_write_fpstate(
-		vcpu: hv_vcpuid_t,
-		buffer: *const c_void,
-		size: size_t,
-	) -> hv_return_t;
+/// Unmaps a region in the guest physical address space of the virutal machine
+pub fn unmap_mem(gpa: u64, size: usize) -> Result<(), Error> {
+	match_error_code(unsafe { hv_vm_unmap(gpa as hv_gpaddr_t, size as size_t) })
 }
 
-// Accessing Machine Specific Registers (MSRs)
-extern "C" {
+/// Synchronizes the guest Timestamp-Counters (TSC) across all VirtualCpus
+///
+/// * `tsc` Guest TSC value
+pub fn sync_tsc(tsc: u64) -> Result<(), Error> {
+	match_error_code(unsafe { hv_vm_sync_tsc(tsc) })
+}
+
+/// Forces an immediate VMEXIT of a set of VirtualCpus
+///
+/// * `VirtualCpu_ids` Array of VirtualCpu IDs
+pub fn interrupt_vcpus(vcpu_ids: &[u32]) -> Result<(), Error> {
+	match_error_code(unsafe { hv_vcpu_interrupt(vcpu_ids.as_ptr(), vcpu_ids.len() as c_uint) })
+}
+
+/// Virtual CPU
+pub struct VirtualCpu {
+	/// Virtual CPU handle
+	id: hv_vcpuid_t,
+}
+
+/// x86 architectural register
+#[derive(Clone)]
+#[repr(C)]
+pub enum Register {
+	RIP,
+	RFLAGS,
+	RAX,
+	RCX,
+	RDX,
+	RBX,
+	RSI,
+	RDI,
+	RSP,
+	RBP,
+	R8,
+	R9,
+	R10,
+	R11,
+	R12,
+	R13,
+	R14,
+	R15,
+	CS,
+	SS,
+	DS,
+	ES,
+	FS,
+	GS,
+	IDT_BASE,
+	IDT_LIMIT,
+	GDT_BASE,
+	GDT_LIMIT,
+	LDTR,
+	LDT_BASE,
+	LDT_LIMIT,
+	LDT_AR,
+	TR,
+	TSS_BASE,
+	TSS_LIMIT,
+	TSS_AR,
+	CR0,
+	CR1,
+	CR2,
+	CR3,
+	CR4,
+	DR0,
+	DR1,
+	DR2,
+	DR3,
+	DR4,
+	DR5,
+	DR6,
+	DR7,
+	TPR,
+	XCR0,
+	REGISTERS_MAX,
+}
+
+impl VirtualCpu {
+	/// Creates a VirtualCpu instance for the current thread
+	pub fn new() -> Result<VirtualCpu, Error> {
+		let mut vcpuid: hv_vcpuid_t = 0;
+
+		match_error_code(unsafe { hv_vcpu_create(&mut vcpuid, HV_VCPU_DEFAULT) })?;
+
+		Ok(VirtualCpu { id: vcpuid })
+	}
+
+	pub fn get_id(&self) -> hv_vcpuid_t {
+		self.id
+	}
+
+	/// Forces an immediate VMEXIT of the VirtualCpu
+	pub fn interrupt(&self) -> Result<(), Error> {
+		match_error_code(unsafe { hv_vcpu_interrupt(&(self.id), 1 as c_uint) })
+	}
+
+	/// Returns the cumulative execution time of the VirtualCpu in nanoseconds
+	pub fn exec_time(&self) -> Result<u64, Error> {
+		let mut exec_time: u64 = 0;
+
+		let _error = match_error_code(unsafe { hv_vcpu_get_exec_time(self.id, &mut exec_time) })?;
+
+		Ok(exec_time)
+	}
+
+	/// Forces flushing of cached VirtualCpu state
+	pub fn flush(&self) -> Result<(), Error> {
+		match_error_code(unsafe { hv_vcpu_flush(self.id) })
+	}
+
+	/// Invalidates the translation lookaside buffer (TLB) of the VirtualCpu
+	pub fn invalidate_tlb(&self) -> Result<(), Error> {
+		match_error_code(unsafe { hv_vcpu_invalidate_tlb(self.id) })
+	}
+
 	/// Enables an MSR to be used natively by the VM
-	pub fn hv_vcpu_enable_native_msr(vcpu: hv_vcpuid_t, msr: u32, enable: bool) -> hv_return_t;
+	pub fn enable_native_msr(&self, msr: u32, enable: bool) -> Result<(), Error> {
+		match_error_code(unsafe { hv_vcpu_enable_native_msr(self.id, msr, enable) })
+	}
 
-	/// Returns the current value of an MSR of a vCPU
-	pub fn hv_vcpu_read_msr(vcpu: hv_vcpuid_t, msr: u32, value: *mut u64) -> hv_return_t;
+	/// Returns the current value of an MSR of the VirtualCpu
+	pub fn read_msr(&self, msr: u32) -> Result<u64, Error> {
+		let mut value: u64 = 0;
 
-	/// Set the value of an MSR of a vCPU
-	pub fn hv_vcpu_write_msr(vcpu: hv_vcpuid_t, msr: u32, value: *const u64) -> hv_return_t;
-}
+		let _error = match_error_code(unsafe { hv_vcpu_read_msr(self.id, msr, &mut value) })?;
 
-// Managing Timestamp-Counters (TSC)
-extern "C" {
-	/// Synchronizes guest Timestamp-Counters (TSC) across all vCPUs
-	pub fn hv_vm_sync_tsc(tsc: u64) -> hv_return_t;
-}
+		Ok(value)
+	}
 
-/// Type of a user virtual address
-pub type hv_uvaddr_t = *const c_void;
+	/// Set the value of an MSR of the VirtualCpu
+	pub fn write_msr(&self, msr: u32, value: u64) -> Result<(), Error> {
+		match_error_code(unsafe { hv_vcpu_write_msr(self.id, msr, &(value)) })
+	}
 
-/// Guest physical memory region permissions for hv_vm_map()
-/// and hv_vm_protect()
-pub type hv_memory_flags_t = u64;
+	/// Returns the current value of an architectural x86 register
+	/// of the VirtualCpu
+	pub fn read_register(&self, reg: &Register) -> Result<u64, Error> {
+		let mut value: u64 = 0;
 
-/// Type of a guest physical address
-pub type hv_gpaddr_t = u64;
+		match_error_code(unsafe { hv_vcpu_read_register(self.id, (*reg).clone(), &mut value) })?;
 
-// Guest physical memory region permissions for hv_vm_map() and hv_vm_protect()
-pub const HV_MEMORY_READ: hv_memory_flags_t = 1 << 0;
-pub const HV_MEMORY_WRITE: hv_memory_flags_t = 1 << 1;
-pub const HV_MEMORY_EXEC: hv_memory_flags_t = 1 << 2;
+		Ok(value)
+	}
 
-// Managing Memory Regions
-extern "C" {
-	/// Maps a region in the virtual address space of the current
-	/// task into the guest physical address space of the VM
-	pub fn hv_vm_map(
-		uva: hv_uvaddr_t,
-		gpa: hv_gpaddr_t,
-		size: size_t,
-		flags: hv_memory_flags_t,
-	) -> hv_return_t;
+	/// Sets the value of an architectural x86 register of the VirtualCpu
+	pub fn write_register(&self, reg: &Register, value: u64) -> Result<(), Error> {
+		match_error_code(unsafe { hv_vcpu_write_register(self.id, (*reg).clone(), value) })
+	}
 
-	/// Unmaps a region in the guest physical address space of the VM
-	pub fn hv_vm_unmap(gpa: hv_gpaddr_t, size: size_t) -> hv_return_t;
+	/// Sets the value of a VMCS field of the VirtualCpu
+	pub fn write_vmcs(&self, field: u32, value: u64) -> Result<(), Error> {
+		match_error_code(unsafe { hv_vmx_vcpu_write_vmcs(self.id, field, value) })
+	}
 
-	/// Modifies the permissions of a region in the guest physical
-	/// address space of the VM
-	pub fn hv_vm_protect(gpa: hv_gpaddr_t, size: size_t, flags: hv_memory_flags_t) -> hv_return_t;
-}
-
-// Managing Virtual Machine Control Structure (VMCS)
-extern "C" {
-	/// Returns the current value of a VMCS field of a vCPU
-	pub fn hv_vmx_vcpu_read_vmcs(vcpu: hv_vcpuid_t, field: u32, value: *mut u64) -> hv_return_t;
-
-	/// Sets the value of a VMCS field of a vCPU
-	pub fn hv_vmx_vcpu_write_vmcs(vcpu: hv_vcpuid_t, field: u32, value: u64) -> hv_return_t;
-
-	/// Returns the VMX capabilities of the host processor
-	pub fn hv_vmx_read_capability(field: super::VMXCap, value: *mut u64) -> hv_return_t;
-
-	/// Sets the address of the guest APIC for a vCPU in the
+	/// Sets the address of the guest APIC for the VirtualCpu in the
 	/// guest physical address space of the VM
-	pub fn hv_vmx_vcpu_set_apic_address(vcpu: hv_vcpuid_t, gpa: hv_gpaddr_t) -> hv_return_t;
+	pub fn set_apic_addr(&self, gpa: u64) -> Result<(), Error> {
+		match_error_code(unsafe { hv_vmx_vcpu_set_apic_address(self.id, gpa) })
+	}
+
+	/// Reads the current architectural x86 floating point and SIMD state of the VirtualCpu
+	pub fn read_fpstate(&self, buffer: &mut [u8]) -> Result<(), Error> {
+		match_error_code(unsafe {
+			hv_vcpu_read_fpstate(
+				self.id,
+				buffer.as_mut_ptr() as *mut c_void,
+				buffer.len() as size_t,
+			)
+		})
+	}
+
+	/// Sets the architectural x86 floating point and SIMD state of the VirtualCpu
+	pub fn write_fpstate(&self, buffer: &[u8]) -> Result<(), Error> {
+		match_error_code(unsafe {
+			hv_vcpu_write_fpstate(
+				self.id,
+				buffer.as_ptr() as *const c_void,
+				buffer.len() as size_t,
+			)
+		})
+	}
+}
+
+/// VMX cabability
+#[allow(non_camel_case_types)]
+#[derive(Clone, Debug)]
+#[repr(C)]
+pub enum VMXCap {
+	/// Pin-based VMX capabilities
+	PINBASED = 0,
+	/// Primary proc-based VMX capabilities
+	PROCBASED = 1,
+	/// Secondary proc-based VMX capabilities
+	PROCBASED2 = 2,
+	/// VM-entry VMX capabilities
+	ENTRY = 3,
+	/// VM-exit VMX capabilities
+	EXIT = 4,
+	/// VMX preemption timer frequency
+	PREEMPTION_TIMER = 32,
+}
+
+/// Reads a VMX capability of the host processor
+pub fn read_vmx_cap(vmx_cap: &VMXCap) -> Result<u64, Error> {
+	let mut value: u64 = 0;
+
+	match_error_code(unsafe { hv_vmx_read_capability((*vmx_cap).clone(), &mut value) })?;
+
+	Ok(value)
+}
+
+impl fmt::Display for VMXCap {
+	fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+		match *self {
+			VMXCap::PINBASED => write!(f, "Pin-based VMX capabilities"),
+			VMXCap::PROCBASED => write!(f, "Primary proc-based VMX capabilities"),
+			VMXCap::PROCBASED2 => write!(f, "Secondary proc-based VMX capabilities"),
+			VMXCap::ENTRY => write!(f, "VM-entry VMX capabilities"),
+			VMXCap::EXIT => write!(f, "VM-exit VMX capabilities"),
+			VMXCap::PREEMPTION_TIMER => write!(f, "VMX preemption timer frequency"),
+		}
+	}
 }
